11sep25

+++ 2 семестр/новый автотест/task100/100.c
@@ -2,36 +2,44 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-int main(void) {
+int main(void)
+{
     int M, N, L, K;
     FILE *in = fopen("data.txt", "r");
     FILE *out = fopen("res.txt", "w");
 
-    if (in == NULL || out == NULL) {
+    if (in == NULL || out == NULL)
+    {
         if (out)
             fprintf(out, "ERROR\n");
-        if (in) fclose(in);
-        if (out) fclose(out);
+        if (in)
+            fclose(in);
+        if (out)
+            fclose(out);
         return -1;
     }
 
-    if (fscanf(in, "%d %d %d %d", &M, &N, &L, &K) != 4 || M <= 0 || L <= 0 || K <= 0) {
+    if (fscanf(in, "%d %d %d %d", &M, &N, &L, &K) != 4 || M <= 0 || L <= 0 || K <= 0)
+    {
         fprintf(out, "ERROR\n");
         fclose(in);
         fclose(out);
         return -1;
     }
 
-    int *A = (int*)malloc(L * K * sizeof(double));
-    if (!A) {
+    int *A = (int *)malloc(L * K * sizeof(double));
+    if (!A)
+    {
         fprintf(out, "ERROR\n");
         fclose(in);
         fclose(out);
         return -1;
     }
 
-    for (int i = 0; i < L * K; i++) {
-        if (fscanf(in, "%d", &A[i]) != 1) {
+    for (int i = 0; i < L * K; i++)
+    {
+        if (fscanf(in, "%d", &A[i]) != 1)
+        {
             fprintf(out, "ERROR\n");
             free(A);
             fclose(in);
@@ -44,7 +52,8 @@ int main(void) {
     // Флаги для строк и столбцов
     int *rows_with_mod = (int *)malloc(L * sizeof(int));
     int *cols_with_mod = (int *)malloc(K * sizeof(int));
-    if (!rows_with_mod || !cols_with_mod) {
+    if (!rows_with_mod || !cols_with_mod)
+    {
         fprintf(out, "ERROR\n");
         free(A);
         free(rows_with_mod);
@@ -54,10 +63,13 @@ int main(void) {
     }
 
     // Отметить строки и столбцы, содержащие элементы с остатком N по модулю M
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
-            int val = (int)A[i * K + j];
-            if (val % M == N) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
+            int val = A[i * K + j];
+            if (val % M == N)
+            {
                 rows_with_mod[i] = 1;
                 cols_with_mod[j] = 1;
             }
@@ -65,18 +77,24 @@ int main(void) {
     }
 
     int *Acopy = (int *)malloc(L * K * sizeof(int));
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
             Acopy[i * K + j] = A[i * K + j];
         }
     }
 
     // Модификация элементов по условиям задачи
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
-            if (rows_with_mod[i] && cols_with_mod[j]) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
+            if (rows_with_mod[i] && cols_with_mod[j])
+            {
                 int count = 0;
-                for (int curi = 0; curi < L; curi++){
+                for (int curi = 0; curi < L; curi++)
+                {
                     if (Acopy[curi * K + j] == Acopy[i * K + j])
                         count++;
                 }
@@ -90,8 +108,10 @@ int main(void) {
 
     // Вывод результатов
     fprintf(out, "%d %d %d %d\n", M, N, L, K);
-    for (int i = 0; i < L; i++) {
-        for (int j = 0; j < K; j++) {
+    for (int i = 0; i < L; i++)
+    {
+        for (int j = 0; j < K; j++)
+        {
             fprintf(out, "%d ", A[i * K + j]);
         }
         fprintf(out, "\n");
